+++ app/package.json
@@ -88,6 +88,7 @@
     "react-hook-form": "^7.51.2",
     "react-image-crop": "^11.0.5",
     "satori": "^0.10.13",
+    "server-only": "^0.0.1",
     "sonner": "^1.4.41",
     "tailwind-merge": "^2.2.2",
     "tailwindcss-animate": "^1.0.7",
